Data_Cleaning_Process
This README file documents the steps taken to clean and standardize the mentor-mentee sessions dataset. The process includes data exploration, removal of duplicates, handling of missing values, standardization of names and dates, and the elimination of irrelevant columns. This cleaned dataset is now optimized for accurate and effective analysis, ensuring data integrity and consistency.